fix positions accrual error and optimize storage slots

+++ src/core/ReentrancyGuard.sol
@@ -6,7 +6,7 @@ pragma solidity >=0.8.0;
 /// @author Modified from OpenZeppelin
 /// (https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/security/ReentrancyGuard.sol)
 abstract contract ReentrancyGuard {
-    uint256 private locked = 1;
+    uint16 private locked = 1;
 
     modifier nonReentrant() virtual {
         require(locked == 1, "REENTRANCY");
